backend/requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
bcrypt
passlib
google-genai
Pillow
pytesseract
sqlalchemy
pydantic
python-dotenv
